requirements.txt
# Backend requirements
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LangChain
langchain==0.0.340
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-chroma==0.1.0

# OpenAI
openai==1.3.0
tiktoken==0.5.1

# Vector DB
chromadb==0.4.22

# Enhanced PDF processing
pymupdf==1.23.8
unstructured==0.10.30
unstructured[pdf]==0.10.30

# Enhanced retrieval
rank_bm25==0.2.2
sentence-transformers==2.2.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
tqdm==4.66.1


# Frontend requirements
streamlit>=1.28.0
requests>=2.31.0
